requirements.txt
google-genai>=1.0.0
elevenlabs>=2.0.0
requests>=2.28.0
